+++ static/css/base/reset.css
@@ -1,0 +1,6 @@
+/* Basic reset */
+*, *::before, *::after {
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+}
