+++ static/css/base/utilities.css
@@ -1,0 +1,7 @@
+/* Utility classes */
+.text-primary {
+    color: var(--primary-color) !important;
+}
+.bg-primary {
+    background-color: var(--primary-color) !important;
+}
