+++ static/css/base/variables.css
@@ -1,0 +1,8 @@
+:root {
+    --primary-color: #2c3e50;
+    --bs-primary: #2c3e50;
+    --secondary-color: #3498db;
+    --accent-color: #2ecc71;
+    --light-bg: #f8f9fa;
+    --dark-bg: #2c3e50;
+}
