+++ static/css/layouts/auth.css
@@ -1,0 +1,7 @@
+/* Styles for auth pages */
+.auth-page {
+    min-height: 100vh;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
