+++ static/css/main.css
@@ -1,0 +1,8 @@
+@import url('base/reset.css');
+@import url('base/variables.css');
+@import url('base/utilities.css');
+@import url('components/header.css');
+@import url('components/sidebar.css');
+@import url('components/messages.css');
+@import url('layouts/dashboard.css');
+@import url('layouts/auth.css');
