+++ static/js/core/app.js
@@ -1,0 +1,3 @@
+import '../components/messages.js';
+import '../components/sidebar.js';
+import '../components/mobileHeader.js';
