+++ static/js/core/utils.js
@@ -1,0 +1,3 @@
+export function qs(selector, scope = document) {
+    return scope.querySelector(selector);
+}
