+++ static/js/pages/settings.js
