+++ templates/auth/login.html
@@ -8,8 +8,6 @@
 {% block page_actions %}
 {% endblock %}
 
-{%block messages%}
-{%endblock%}
 
 {%block css%}
 <style>
@@ -92,15 +90,7 @@
             <p>Acesso ao Sistema</p>
         </div>
         
-        <!-- Alertas -->
-        {% if messages %}
-            {% for message in messages %}
-                <div class="alert alert-warning alert-dismissible fade show" role="alert">
-                    {{ message }}
-                    <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
-                </div>
-            {% endfor %}
-        {% endif %}
+        {% include 'components/messages.html' %}
         
         <form method="post" action="{% url 'login' %}" class="login-form">
             {% csrf_token %}
