+++ templates/auth/register.html
@@ -8,8 +8,6 @@
 {% block page_actions %}
 {% endblock %}
 
-{%block messages%}
-{%endblock%}
 
 {%block css%}
 <style>
@@ -92,15 +90,7 @@
             <p>Cadastro de Usuário</p>
         </div>
         
-        <!-- Alertas -->
-        {% if messages %}
-            {% for message in messages %}
-                <div class="alert alert-{{ message.tags }} alert-dismissible fade show" role="alert">
-                    {{ message }}
-                    <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
-                </div>
-            {% endfor %}
-        {% endif %}
+        {% include 'components/messages.html' %}
         
         <form method="post" action="" class="register-form">
             {% csrf_token %}
