+++ templates/components/messages.html
@@ -1,10 +1,10 @@
 {% if messages %}
-        <div class="messages-container position-fixed top-0 end-0 p-3 mt-2 me-2" style="z-index: 1050; max-width: 300px;">
-            {% for message in messages %}
-            <div class="alert alert-{{ message.tags }} shadow-sm mb-2 p-2 ps-3 animate__animated animate__fadeInRight">
-                <small>{{ message }}</small>
-                <button type="button" class="btn-close btn-close-sm float-end ms-2" data-bs-dismiss="alert" aria-label="Close"></button>
-            </div>
-            {% endfor %}
-        </div>
-    {% endif %}
+<div class="messages-container position-fixed top-0 end-0 p-3 mt-2 me-2" style="z-index: 1050; max-width: 300px;">
+    {% for message in messages %}
+    <div class="alert alert-{{ message.tags }} shadow-sm mb-2 p-2 ps-3 animate__animated animate__fadeInRight">
+        <small>{{ message }}</small>
+        <button type="button" class="btn-close btn-close-sm float-end ms-2" data-bs-dismiss="alert" aria-label="Close"></button>
+    </div>
+    {% endfor %}
+</div>
+{% endif %}
