+++ templates/default.html
@@ -7,37 +7,14 @@
     <title>{{config.SITE_NAME}}{%block title%}{%endblock%}</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
-    <link rel="stylesheet" href="{% static 'css/global.css' %}">
-    <link rel="stylesheet" href="{% static 'css/sidebar.css' %}">
-    <link rel="stylesheet" href="{% static 'css/mobile-header.css' %}">
-    <style>
-        :root {
-            --primary-color: #2c3e50;
-            --bs-primary: #2c3e50;
-            --secondary-color: #3498db;
-            --accent-color: #2ecc71;
-            --light-bg: #f8f9fa;
-            --dark-bg: #2c3e50;
-        }
-    </style>
+    <link rel="stylesheet" href="{% static 'css/main.css' %}">
     {%block css%}{%endblock%}
 </head>
 <body>
     <!-- Sistema de mensagens -->
-    {%block messages%}
-        {% if messages %}
-        <div class="messages-container position-fixed top-0 end-0 p-3 mt-2 me-2" style="z-index: 1050; max-width: 300px;">
-            {% for message in messages %}
-            {% if message.tags == 'error' %}
-                <div class="alert alert-danger shadow-sm mb-2 p-2 ps-3 animate__animated animate__fadeInRight">
-            {% endif %}
-                <small>{{ message }}</small>
-                <!-- <button type="button" class="btn-close btn-close-sm float-end ms-2" data-bs-dismiss="alert" aria-label="Close"></button> -->
-            </div>
-            {% endfor %}
-        </div>
-        {% endif %}
-    {%endblock%}
+    {% block messages %}
+        {% include 'components/messages.html' %}
+    {% endblock %}
     
     <!-- Mobile Header (new) -->
     {% include 'components/mobile-header.html' %}
@@ -85,9 +62,7 @@
     <script src="https://code.jquery.com/jquery-3.7.1.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"></script>
     
-    <script src="{% static 'js/globalController.js' %}"></script>
-    <script src="{% static 'js/app/sidebarController.js' %}"></script>
-    <script src="{% static 'js/mobile-header.js' %}"></script>
+    <script type="module" src="{% static 'js/core/app.js' %}"></script>
     
     {% block js %}
     {% endblock %}
